requirements.txt
streamlit
requests
pandas
numpy
python-dotenv
folium